% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicor.R
\name{multicor}
\alias{multicor}
\title{Caculate the multiple correlation}
\usage{
multicor(data, method = c("pearson", "kendall", "spearman"), rds = NA,
  rewrite = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{expression matrix}

\item{method}{a character string indicating which correlation coefficient (or covariance) is to be computed.}

\item{rds}{RDS to save data}

\item{rewrite}{rewrite RDS}

\item{verbose}{log level}
}
\value{
v1, v2, r, p.value, FDR
}
\description{
Caculate the multiple correlation of expression matrix. The matrix is
like this:
     sample1  sample2  sample3
gene1   xx      xx      xx
gene2   xx      xx      xx
gene3   xx      xx      xx
}
\author{
Navy Cheng
}
