% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prad.rna.count}
\alias{prad.rna.count}
\title{The RNA-seq counts data from TCGA-PRAD project}
\format{A data frame with 60483 rows and 551 samples.}
\usage{
prad.rna.count
}
\description{
The counts data contain 551 samples with 499 tumor and 52 normal tissue.
}
\keyword{datasets}
